[AIRFLOW-6721] Move NamedHivePartitionSensor tests
Move the NamedHivePartitionSensor tests from the operators module into
the test module for NamedHivePartitionSensor.
